requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# ControlNet and Image Processing
controlnet-aux>=0.0.7
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# UI Frameworks
streamlit>=1.28.0
gradio>=4.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
